requirements.txt
socialoauth
redis
django-celery
